apartment_monitor/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
schedule>=1.2.0
twilio>=8.2.0
python-dotenv>=1.0.0